contract_bot/requirements.txt
streamlit
spacy
pdfplumber
python-docx
reportlab
openai
nltk
